COM-1467 - rename folder
