+++ src/components/design/AnimatedShadow.tsx
