+++ src/components/design/FooterGradient.tsx
